# openapi.yaml
openapi: 3.0.0
info:
  title: Blockchain Indexer API
  version: 1.0.0
  description: API for processing blocks, getting balances, and rolling back the blockchain state.
servers:
  - url: http://localhost:3000
paths:
...
  /balance/{address}:
    get:
      summary: Get balance for an address
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Balance retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: number
        "404":
          description: Address not found
...